rumqttc

Module mqttbytes

mqttbytes

This module contains the low level struct definitions required to assemble and disassemble MQTT 3.1.1 packets in rumqttc. The bytes crate is used internally.

Modules

v4

Structs

FixedHeader

Packet type from a byte

Enums

Error

Error during serialization and deserialization

PacketType

MQTT packet type

Protocol

Protocol type

QoS

Quality of service

Functions

check

Checks if the stream has enough bytes to frame a packet and returns fixed header only if a packet can be framed with existing bytes in the stream. The passed stream doesn’t modify parent stream’s cursor. If this function returned an error, next check on the same parent stream is forced start with cursor at 0 again (Iter is owned. Only Iter’s cursor is changed internally)

has_wildcards

Checks if a topic or topic filter has wildcards

matches

Checks if topic matches a filter. topic and filter validation isn’t done here.

qos

Maps a number to QoS

valid_filter

Checks if the filter is valid

valid_topic

Checks if a topic is valid