rumqttc::mqttbytes

Module v4

Structs

ConnAck

Acknowledgement to connect packet

Connect

Connection packet initiated by the client

Disconnect

LastWill

LastWill that broker forwards on behalf of the client

Login

PingReq

PingResp

PubAck

Acknowledgement to QoS1 publish

PubComp

QoS2 Assured publish complete, in response to PUBREL packet

PubRec

Acknowledgement to QoS2 publish

PubRel

QoS2 Publish release, in response to PUBREC packet

Publish

Publish packet

SubAck

Acknowledgement to subscribe

Subscribe

Subscription packet

SubscribeFilter

Subscription filter

UnsubAck

Acknowledgement to unsubscribe

Unsubscribe

Unsubscribe packet

Enums

ConnectReturnCode

Return code in connack

Packet

Encapsulates all MQTT packet types

RetainForwardRule

SubscribeReasonCode

Functions

read

Reads a stream of bytes and extracts next MQTT packet out of it