rumqttc

Struct MqttOptions

pub struct MqttOptions { /* private fields */ }

Options to configure the behaviour of MQTT connection

Implementations

impl MqttOptions

pub fn new<S: Into<String>, T: Into<String>>( id: S, host: T, port: u16, ) -> MqttOptions

Create an MqttOptions object that contains default values for all settings other than

id: A string to identify the device connecting to a broker

host: The broker’s domain name or IP address

port: The port number on which broker must be listening for incoming connections

let options = MqttOptions::new("123", "localhost", 1883);

pub fn broker_address(&self) -> (String, u16)

Broker address

pub fn set_last_will(&mut self, will: LastWill) -> &mut Self

pub fn last_will(&self) -> Option<LastWill>

pub fn set_transport(&mut self, transport: Transport) -> &mut Self

pub fn transport(&self) -> Transport

pub fn set_keep_alive(&mut self, duration: Duration) -> &mut Self

Set number of seconds after which client should ping the broker if there is no other data exchange

pub fn keep_alive(&self) -> Duration

Keep alive time

pub fn client_id(&self) -> String

Client identifier

pub fn set_max_packet_size( &mut self, incoming: usize, outgoing: usize, ) -> &mut Self

Set packet size limit for outgoing an incoming packets

pub fn max_packet_size(&self) -> usize

Maximum packet size

pub fn set_clean_session(&mut self, clean_session: bool) -> &mut Self

clean_session = true removes all the state from queues & instructs the broker to clean all the client state when client disconnects.

When set false, broker will hold the client state and performs pending operations on the client when reconnection with same client_id happens. Local queue state is also held to retransmit packets after reconnection.

Panic

Panics if clean_session is false when client_id is empty.

let mut options = MqttOptions::new("", "localhost", 1883);
options.set_clean_session(false);

pub fn clean_session(&self) -> bool

Clean session

pub fn set_credentials<U: Into<String>, P: Into<String>>( &mut self, username: U, password: P, ) -> &mut Self

Username and password

pub fn credentials(&self) -> Option<(String, String)>

Security options

pub fn set_request_channel_capacity(&mut self, capacity: usize) -> &mut Self

Set request channel capacity

pub fn request_channel_capacity(&self) -> usize

Request channel capacity

pub fn set_pending_throttle(&mut self, duration: Duration) -> &mut Self

Enables throttling and sets outoing message rate to the specified ‘rate’

pub fn pending_throttle(&self) -> Duration

Outgoing message rate

pub fn set_inflight(&mut self, inflight: u16) -> &mut Self

Set number of concurrent in flight messages

pub fn inflight(&self) -> u16

Number of concurrent in flight messages

pub fn set_manual_acks(&mut self, manual_acks: bool) -> &mut Self

set manual acknowledgements

pub fn manual_acks(&self) -> bool

get manual acknowledgements