rumqttc::v5

Module mqttbytes

Modules

v5

This module is the place where all the protocol specifics gets abstracted out and creates a structures which are common across protocols. Since, MQTT is the core protocol that this broker supports, a lot of structs closely map to what MQTT specifies in its protocol

Enums

Error

Error during serialization and deserialization

QoS

Quality of service

Functions

has_wildcards

Checks if a topic or topic filter has wildcards

matches

Checks if topic matches a filter. topic and filter validation isn’t done here.

qos

Maps a number to QoS

valid_filter

Checks if the filter is valid

valid_topic

Checks if a topic is valid