rumqttc::v5::mqttbytes::v5

Enum DisconnectReasonCode

#[repr(u8)]
pub enum DisconnectReasonCode {

    NormalDisconnection = 0,
    DisconnectWithWillMessage = 4,
    UnspecifiedError = 128,
    MalformedPacket = 129,
    ProtocolError = 130,
    ImplementationSpecificError = 131,
    NotAuthorized = 135,
    ServerBusy = 137,
    ServerShuttingDown = 139,
    KeepAliveTimeout = 141,
    SessionTakenOver = 142,
    TopicFilterInvalid = 143,
    TopicNameInvalid = 144,
    ReceiveMaximumExceeded = 147,
    TopicAliasInvalid = 148,
    PacketTooLarge = 149,
    MessageRateTooHigh = 150,
    QuotaExceeded = 151,
    AdministrativeAction = 152,
    PayloadFormatInvalid = 153,
    RetainNotSupported = 154,
    QoSNotSupported = 155,
    UseAnotherServer = 156,
    ServerMoved = 157,
    SharedSubscriptionNotSupported = 158,
    ConnectionRateExceeded = 159,
    MaximumConnectTime = 160,
    SubscriptionIdentifiersNotSupported = 161,
    WildcardSubscriptionsNotSupported = 162,

}

Variants

NormalDisconnection = 0

Close the connection normally. Do not send the Will Message.

DisconnectWithWillMessage = 4

The Client wishes to disconnect but requires that the Server also publishes its Will Message.

UnspecifiedError = 128

The Connection is closed but the sender either does not wish to reveal the reason, or none of the other Reason Codes apply.

MalformedPacket = 129

The received packet does not conform to this specification.

ProtocolError = 130

An unexpected or out of order packet was received.

ImplementationSpecificError = 131

The packet received is valid but cannot be processed by this implementation.

NotAuthorized = 135

The request is not authorized.

ServerBusy = 137

The Server is busy and cannot continue processing requests from this Client.

ServerShuttingDown = 139

The Server is shutting down.

KeepAliveTimeout = 141

The Connection is closed because no packet has been received for 1.5 times the Keepalive time.

SessionTakenOver = 142

Another Connection using the same ClientID has connected causing this Connection to be closed.

TopicFilterInvalid = 143

The Topic Filter is correctly formed, but is not accepted by this Sever.

TopicNameInvalid = 144

The Topic Name is correctly formed, but is not accepted by this Client or Server.

ReceiveMaximumExceeded = 147

The Client or Server has received more than Receive Maximum publication for which it has not sent PUBACK or PUBCOMP.

TopicAliasInvalid = 148

The Client or Server has received a PUBLISH packet containing a Topic Alias which is greater than the Maximum Topic Alias it sent in the CONNECT or CONNACK packet.

PacketTooLarge = 149

The packet size is greater than Maximum Packet Size for this Client or Server.

MessageRateTooHigh = 150

The received data rate is too high.

QuotaExceeded = 151

An implementation or administrative imposed limit has been exceeded.

AdministrativeAction = 152

The Connection is closed due to an administrative action.

PayloadFormatInvalid = 153

The payload format does not match the one specified by the Payload Format Indicator.

RetainNotSupported = 154

The Server has does not support retained messages.

QoSNotSupported = 155

The Client specified a QoS greater than the QoS specified in a Maximum QoS in the CONNACK.

UseAnotherServer = 156

The Client should temporarily change its Server.

ServerMoved = 157

The Server is moved and the Client should permanently change its server location.

SharedSubscriptionNotSupported = 158

The Server does not support Shared Subscriptions.

ConnectionRateExceeded = 159

This connection is closed because the connection rate is too high.

MaximumConnectTime = 160

The maximum connection time authorized for this connection has been exceeded.

SubscriptionIdentifiersNotSupported = 161

The Server does not support Subscription Identifiers; the subscription is not accepted.

WildcardSubscriptionsNotSupported = 162

The Server does not support Wildcard subscription; the subscription is not accepted.