rumqttc::v5

Struct Client

pub struct Client { /* private fields */ }

A synchronous client, communicates with MQTT EventLoop.

This is cloneable and can be used to synchronously publish, subscribe through the EventLoop/Connection, which is to be polled in parallel by iterating over the object returned by Connection.iter() in a separate thread.

NOTE: The EventLoop/Connection must be regularly polled(.next() in case of Connection) in order to send, receive and process packets from the broker, i.e. move ahead.

An asynchronous channel handle can also be extracted if necessary.

Implementations

impl Client

pub fn new(options: MqttOptions, cap: usize) -> (Client, Connection)

Create a new Client

cap specifies the capacity of the bounded async channel.

pub fn from_sender(request_tx: Sender<Request>) -> Client

Create a new Client from a channel Sender.

This is mostly useful for creating a test instance where you can listen on the corresponding receiver.

pub fn publish_with_properties<S, P>( &self, topic: S, qos: QoS, retain: bool, payload: P, properties: PublishProperties, ) -> Result<(), ClientError>
where S: Into<String>, P: Into<Bytes>,

pub fn publish<S, P>( &self, topic: S, qos: QoS, retain: bool, payload: P, ) -> Result<(), ClientError>
where S: Into<String>, P: Into<Bytes>,

pub fn try_publish_with_properties<S, P>( &self, topic: S, qos: QoS, retain: bool, payload: P, properties: PublishProperties, ) -> Result<(), ClientError>
where S: Into<String>, P: Into<Bytes>,

pub fn try_publish<S, P>( &self, topic: S, qos: QoS, retain: bool, payload: P, ) -> Result<(), ClientError>
where S: Into<String>, P: Into<Bytes>,

pub fn ack(&self, publish: &Publish) -> Result<(), ClientError>

Sends a MQTT PubAck to the EventLoop. Only needed in if manual_acks flag is set.

pub fn try_ack(&self, publish: &Publish) -> Result<(), ClientError>

Sends a MQTT PubAck to the EventLoop. Only needed in if manual_acks flag is set.

pub fn subscribe_with_properties<S: Into<String>>( &self, topic: S, qos: QoS, properties: SubscribeProperties, ) -> Result<(), ClientError>

pub fn subscribe<S: Into<String>>( &self, topic: S, qos: QoS, ) -> Result<(), ClientError>

pub fn try_subscribe_with_properties<S: Into<String>>( &self, topic: S, qos: QoS, properties: SubscribeProperties, ) -> Result<(), ClientError>

Sends a MQTT Subscribe to the EventLoop

pub fn try_subscribe<S: Into<String>>( &self, topic: S, qos: QoS, ) -> Result<(), ClientError>

pub fn subscribe_many_with_properties<T>( &self, topics: T, properties: SubscribeProperties, ) -> Result<(), ClientError>
where T: IntoIterator<Item = Filter>,

pub fn subscribe_many<T>(&self, topics: T) -> Result<(), ClientError>
where T: IntoIterator<Item = Filter>,

pub fn try_subscribe_many_with_properties<T>( &self, topics: T, properties: SubscribeProperties, ) -> Result<(), ClientError>
where T: IntoIterator<Item = Filter>,

pub fn try_subscribe_many<T>(&self, topics: T) -> Result<(), ClientError>
where T: IntoIterator<Item = Filter>,

pub fn unsubscribe_with_properties<S: Into<String>>( &self, topic: S, properties: UnsubscribeProperties, ) -> Result<(), ClientError>

pub fn unsubscribe<S: Into<String>>(&self, topic: S) -> Result<(), ClientError>

pub fn try_unsubscribe_with_properties<S: Into<String>>( &self, topic: S, properties: UnsubscribeProperties, ) -> Result<(), ClientError>

Sends a MQTT Unsubscribe to the EventLoop

pub fn try_unsubscribe<S: Into<String>>( &self, topic: S, ) -> Result<(), ClientError>

pub fn disconnect(&self) -> Result<(), ClientError>

Sends a MQTT disconnect to the EventLoop

pub fn try_disconnect(&self) -> Result<(), ClientError>

Sends a MQTT disconnect to the EventLoop