rumqttc::v5

Struct MqttOptions

pub struct MqttOptions { /* private fields */ }

Options to configure the behaviour of MQTT connection

Implementations

impl MqttOptions

pub fn new<S: Into<String>, T: Into<String>>( id: S, host: T, port: u16, ) -> MqttOptions

Create an MqttOptions object that contains default values for all settings other than

id: A string to identify the device connecting to a broker

host: The broker’s domain name or IP address

port: The port number on which broker must be listening for incoming connections

let options = MqttOptions::new("123", "localhost", 1883);

pub fn broker_address(&self) -> (String, u16)

Broker address

pub fn set_last_will(&mut self, will: LastWill) -> &mut Self

pub fn last_will(&self) -> Option<LastWill>

pub fn set_transport(&mut self, transport: Transport) -> &mut Self

pub fn transport(&self) -> Transport

pub fn set_keep_alive(&mut self, duration: Duration) -> &mut Self

Set number of seconds after which client should ping the broker if there is no other data exchange

pub fn keep_alive(&self) -> Duration

Keep alive time

pub fn client_id(&self) -> String

Client identifier

pub fn set_clean_start(&mut self, clean_start: bool) -> &mut Self

clean_start = true removes all the state from queues & instructs the broker to clean all the client state when client disconnects.

When set false, broker will hold the client state and performs pending operations on the client when reconnection with same client_id happens. Local queue state is also held to retransmit packets after reconnection.

pub fn clean_start(&self) -> bool

Clean session

pub fn set_credentials<U: Into<String>, P: Into<String>>( &mut self, username: U, password: P, ) -> &mut Self

Username and password

pub fn credentials(&self) -> Option<(String, String)>

Security options

pub fn set_request_channel_capacity(&mut self, capacity: usize) -> &mut Self

Set request channel capacity

pub fn request_channel_capacity(&self) -> usize

Request channel capacity

pub fn set_pending_throttle(&mut self, duration: Duration) -> &mut Self

Enables throttling and sets outoing message rate to the specified ‘rate’

pub fn pending_throttle(&self) -> Duration

Outgoing message rate

pub fn set_connection_timeout(&mut self, timeout: u64) -> &mut Self

set connection timeout in secs

pub fn connection_timeout(&self) -> u64

get timeout in secs

pub fn set_connect_properties( &mut self, properties: ConnectProperties, ) -> &mut Self

set connection properties

pub fn connect_properties(&self) -> Option<ConnectProperties>

get connection properties

pub fn set_receive_maximum(&mut self, recv_max: Option<u16>) -> &mut Self

set receive maximum on connection properties

pub fn receive_maximum(&self) -> Option<u16>

get receive maximum from connection properties

pub fn set_max_packet_size(&mut self, max_size: Option<u32>) -> &mut Self

set max packet size on connection properties

pub fn max_packet_size(&self) -> Option<u32>

get max packet size from connection properties

pub fn set_topic_alias_max(&mut self, topic_alias_max: Option<u16>) -> &mut Self

set max topic alias on connection properties

pub fn topic_alias_max(&self) -> Option<u16>

get max topic alias from connection properties

pub fn set_request_response_info( &mut self, request_response_info: Option<u8>, ) -> &mut Self

set request response info on connection properties

pub fn request_response_info(&self) -> Option<u8>

get request response info from connection properties

pub fn set_request_problem_info( &mut self, request_problem_info: Option<u8>, ) -> &mut Self

set request problem info on connection properties

pub fn request_problem_info(&self) -> Option<u8>

get request problem info from connection properties

pub fn set_user_properties( &mut self, user_properties: Vec<(String, String)>, ) -> &mut Self

set user properties on connection properties

pub fn user_properties(&self) -> Vec<(String, String)>

get user properties from connection properties

pub fn set_authentication_method( &mut self, authentication_method: Option<String>, ) -> &mut Self

set authentication method on connection properties

pub fn authentication_method(&self) -> Option<String>

get authentication method from connection properties

pub fn set_authentication_data( &mut self, authentication_data: Option<Bytes>, ) -> &mut Self

set authentication data on connection properties

pub fn authentication_data(&self) -> Option<Bytes>

get authentication data from connection properties

pub fn set_manual_acks(&mut self, manual_acks: bool) -> &mut Self

set manual acknowledgements

pub fn manual_acks(&self) -> bool

get manual acknowledgements

pub fn network_options(&self) -> NetworkOptions

pub fn set_network_options( &mut self, network_options: NetworkOptions, ) -> &mut Self

pub fn set_outgoing_inflight_upper_limit(&mut self, limit: u16) -> &mut Self

Get the upper limit on maximum number of inflight outgoing publishes. The server may set its own maximum inflight limit, the smaller of the two will be used.

pub fn get_outgoing_inflight_upper_limit(&self) -> Option<u16>

Set the upper limit on maximum number of inflight outgoing publishes. The server may set its own maximum inflight limit, the smaller of the two will be used.